start separate assert mul with the goal of compatibility, to swap out later
